+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added displaying a list of detected unsupported renderers on the React renderer connection waiting screen. This should help figure out why a React can't be connected.
 - Added ancestors section in component's details panel
 - Fixed memory leaks related to keeping unmounted fibers by RRT internal dictionaries
+- Fixed navigation when a fiber is pinned
 - Removed "Blocking mode" mark for a render root as outdated
 - Considered `production` and `profiling` React builds as partially functional, i.e. RRT works but some data is not capturing